requirements.txt
huggingface-hub==0.25.1
python-dotenv==1.0.1
python-telegram-bot==21.6
